work

+++ ft_memmove.c
@@ -6,7 +6,7 @@
 /*   By: mmasima <marvin@42.fr>                     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/06/03 16:45:04 by mmasima           #+#    #+#             */
-/*   Updated: 2019/06/05 14:15:09 by mmasima          ###   ########.fr       */
+/*   Updated: 2019/06/05 14:54:49 by mmasima          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 void	*ft_memmove(void *dst, const void *src, size_t len)
 {
-	int				x;
+	size_t			x;
 	unsigned char	*ptr1;
 	unsigned char	*ptr2;
